+++ client/src/pages/signin/signin.css
@@ -6,46 +6,45 @@
   justify-content: center;
   margin-top: 10px;
   overflow: auto;
+  width: auto;
+  margin: auto;
 }
 
 .wrapper{
-    display: flex;
+  display: flex;
   align-items: center;
   flex-direction: column;
-  background-color: transparent;
-  color: black;
+  width: 300px;
+  height: 350px;
   border: 1px solid black;
-  border-radius: 5px;
-  padding: 20px 50px;
-  gap: 10px;
 }
 h5{
-    font-size: 24px;
-   margin-top: 10px;
-   color: #9A2585;
+    color: #9A2585;
+    text-decoration: none;
+    font-size: 18px;
+    text-align: center;
 }
 
 .input{
-    border: 1px solid black;
-  border-radius: 3px;
-  padding: 10px;
-  margin: 2px;
-  background-color: transparent;
-  width: 100%;
-  color: black;
+  width: 250px;
+  padding: 12px;
   outline: none;
+  margin: 10px;
+  margin-top: 15px;
+  border-radius: 5px;
+  color: black;
 }
 
 .signin-button{
-    border-radius: 3px;
-  border: none;
-  margin-top: 2px;
-  padding: 10px 10px;
-  font-weight: 500;
-  cursor: pointer;
+  padding: 10px;
+  border-radius: 5px;
+  width:250px;
+  font-size: 16px;
   background-color: dodgerblue;
+  cursor: pointer;
+  margin: 10px;
   color: white;
-  width: 60%;
+  margin-bottom: 20px;
 }
 
 .alt_login{
